DraggableScrollableSheet - drag/fling consistency
Make the handling of ballistic and drag consistent in the case of
scrolled inner content. Previously, when dragging upwards, the drag
would scroll the inner scrollable but the ballistic scroll would then
expand the outer sheet.

Makes up for the loss of a way to expand the outer sheet by making both
drag and ballistic scroll the outer sheet if the inner scrollable is at
its end.